requirements.txt
pandas==2.3.3
numpy==2.4.2
scikit-learn==1.8.0
mlflow==3.10.0
fastapi[standard]==0.135.1
skops==0.13.0
azureml-mlflow==1.60.0
